requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0

# Database
PyMySQL>=1.1,<2.0
cryptography>=41.0,<43.0

# Caching & Task Queue
redis>=5.0,<6.0
django-redis>=5.4,<6.0
celery>=5.3,<6.0

# Production Server
gunicorn>=21.0,<22.0

# Development & Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
flake8>=6.1,<7.0
black>=23.0,<24.0

# Utilities
python-dotenv>=1.0,<2.0
